requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
Flask-CORS==4.0.0
python-dotenv==1.0.0
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0